@@ -3,40 +3,40 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/react-hooks": "^7.0.1",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.22",
-    "@types/node": "^14.14.37",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.3",
-    "@types/react-redux": "^7.1.16",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/react-slick": "^0.23.4",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/react-hooks": "^8.0.1",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.5.1",
+    "@types/node": "^20.2.1",
+    "@types/react": "^18.2.6",
+    "@types/react-dom": "^18.2.4",
+    "@types/react-redux": "^7.1.25",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-slick": "^0.23.10",
     "@types/redux-mock-store": "^1.0.3",
-    "@types/styled-components": "^5.1.9",
-    "@types/yup": "^0.29.11",
-    "axios": "^0.21.1",
-    "classnames": "^2.3.0",
-    "formik": "^2.2.6",
+    "@types/styled-components": "^5.1.26",
+    "@types/yup": "^0.29.14",
+    "axios": "^1.4.0",
+    "classnames": "^2.3.2",
+    "formik": "^2.2.9",
     "jest-dom": "^4.0.0",
-    "json-server": "^0.16.3",
+    "json-server": "^0.17.3",
     "jwt-decode": "^3.1.2",
-    "react": "^17.0.1",
-    "react-content-loader": "^6.0.3",
-    "react-dom": "^17.0.1",
-    "react-redux": "^7.2.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-slick": "^0.28.1",
-    "redux": "^4.0.5",
+    "react": "^18.2.0",
+    "react-content-loader": "^6.2.1",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.11.2",
+    "react-scripts": "5.0.1",
+    "react-slick": "^0.29.0",
+    "redux": "^4.2.1",
     "redux-mock-store": "^1.5.4",
-    "redux-thunk": "^2.3.0",
-    "sass": "^1.32.8",
-    "styled-components": "^5.2.1",
-    "typescript": "^4.2.3",
-    "web-vitals": "^1.0.1"
+    "redux-thunk": "^2.4.2",
+    "sass": "^1.62.1",
+    "styled-components": "^5.3.10",
+    "typescript": "^5.0.4",
+    "web-vitals": "^3.3.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -63,6 +63,7 @@
     ]
   },
   "devDependencies": {
-    "prettier": "2.2.1"
+    "@types/rc-slider": "^9.3.1",
+    "prettier": "2.8.8"
   }
 }
