@@ -1,7 +1,7 @@
-import React from 'react';
+import React from "react";
+import styled from "styled-components";
 
-import styled from 'styled-components';
-import Home from './pages/Home';
+import Home from "./pages/Home";
 
 const AppWrapper = styled.div`
   background-color: #fff;
